Move 072 files to common folder and use it in Instant60
